Move Parameter Files to ParamModel
