+++ Griveance/BusinessLayer/GetSingleFaculty.cs
@@ -1,4 +1,5 @@
 ﻿using Griveance.Models;
+using Griveance.ParamModel;
 using Griveance.ResultModel;
 using System;
 using System.Collections.Generic;
