+++ Griveance/ParamModel/ParamUser.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 
-namespace Griveance.Models
+namespace Griveance.ParamModel
 {
     public class ParamUser
     {
